Move the contents of deblend.py to scarletDeblentTask.py
This is primarily to avoid shadowing the deblend function at
lsst.meas.extensions.scarlet.deblend.

The old interface is retained but deprecated for removal after the next major
release.
