+++ python/lsst/meas/extensions/scarlet/__init__.py
@@ -19,8 +19,7 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
-#from .version import *  # Generated by sconsUtils
-
-from . import source
-from . import deblend
 from .version import *
+
+from .source import *
+from .scarletDeblendTask import *
